fix: move withTransition types to a component file
