+++ packages/transition/index.ts
@@ -1,2 +1,2 @@
 export * from './withTransition'
-export * from './types'
+
