+++ packages/transition/withTransition.tsx
@@ -9,8 +9,35 @@ import {
 import { useMergeRefs } from '../hooks'
 import { Transition } from 'react-transition-group'
 import { DEFAULT_DURATION } from './constants'
-import { TransitionWrapperProps, TransitionInnerProps } from './types'
 
+import type {
+  TransitionProps,
+  TransitionStatus,
+  TransitionActions,
+} from 'react-transition-group/Transition'
+
+export type TransitionWrapperProps = Partial<
+  Pick<
+    TransitionProps<HTMLElement>,
+    | 'in'
+    | 'mountOnEnter'
+    | 'unmountOnExit'
+    | 'timeout'
+    | 'addEndListener'
+    | 'onEnter'
+    | 'onEntering'
+    | 'onEntered'
+    | 'onExit'
+    | 'onExiting'
+    | 'onExited'
+  > &
+  Pick<TransitionActions, 'appear' | 'enter' | 'exit'>
+>
+
+export type TransitionInnerProps = {
+  duration: number
+  transitionStatus: TransitionStatus
+}
 type WrappedProps<T> = Omit<T, keyof TransitionInnerProps> &
   TransitionWrapperProps
